move tools to own pkg & add generic network waits
