+++ src/cmd/tools/crane.go
@@ -1,0 +1,79 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2021-Present The Zarf Authors
+
+// Package tools contains the CLI commands for Zarf.
+package tools
+
+import (
+	"github.com/defenseunicorns/zarf/src/config"
+	"github.com/defenseunicorns/zarf/src/config/lang"
+	"github.com/defenseunicorns/zarf/src/internal/cluster"
+	craneCmd "github.com/google/go-containerregistry/cmd/crane/cmd"
+	"github.com/google/go-containerregistry/pkg/crane"
+	"github.com/spf13/cobra"
+)
+
+func init() {
+	cranePlatformOptions := config.GetCraneOptions(false)
+
+	craneLogin := craneCmd.NewCmdAuthLogin()
+	craneLogin.Example = ""
+
+	registryCmd := &cobra.Command{
+		Use:     "registry",
+		Aliases: []string{"r", "crane"},
+		Short:   lang.CmdToolsRegistryShort,
+	}
+
+	toolsCmd.AddCommand(registryCmd)
+
+	registryCmd.AddCommand(craneLogin)
+	registryCmd.AddCommand(craneCmd.NewCmdPull(&cranePlatformOptions))
+	registryCmd.AddCommand(craneCmd.NewCmdPush(&cranePlatformOptions))
+	registryCmd.AddCommand(craneCmd.NewCmdCopy(&cranePlatformOptions))
+	registryCmd.AddCommand(zarfCraneCatalog(&cranePlatformOptions))
+}
+
+// Wrap the original crane catalog with a zarf specific version
+func zarfCraneCatalog(cranePlatformOptions *[]crane.Option) *cobra.Command {
+	craneCatalog := craneCmd.NewCmdCatalog(cranePlatformOptions)
+
+	eg := `  # list the repos internal to Zarf
+  $ zarf tools registry catalog
+
+  # list the repos for reg.example.com
+  $ zarf tools registry catalog reg.example.com`
+
+	craneCatalog.Example = eg
+	craneCatalog.Args = nil
+
+	originalCatalogFn := craneCatalog.RunE
+
+	craneCatalog.RunE = func(cmd *cobra.Command, args []string) error {
+		if len(args) > 0 {
+			return originalCatalogFn(cmd, args)
+		}
+
+		// Load Zarf state
+		zarfState, err := cluster.NewClusterOrDie().LoadZarfState()
+		if err != nil {
+			return err
+		}
+
+		// Open a tunnel to the Zarf registry
+		tunnelReg, err := cluster.NewZarfTunnel()
+		if err != nil {
+			return err
+		}
+		tunnelReg.Connect(cluster.ZarfRegistry, false)
+
+		// Add the correct authentication to the crane command options
+		authOption := config.GetCraneAuthOption(zarfState.RegistryInfo.PullUsername, zarfState.RegistryInfo.PullPassword)
+		*cranePlatformOptions = append(*cranePlatformOptions, authOption)
+		registryEndpoint := tunnelReg.Endpoint()
+
+		return originalCatalogFn(cmd, []string{registryEndpoint})
+	}
+
+	return craneCatalog
+}
