+++ src/config/config.go
@@ -79,6 +79,9 @@ var (
 	// Dirty Solution to getting the real time deployedComponents components.
 	deployedComponents []types.DeployedComponent
 
+	// SkipLogFile is a flag to skip logging to a file
+	SkipLogFile bool
+
 	SGetPublicKey string
 	UIAssets      embed.FS
